+++ js/modules/utils.js
@@ -1,0 +1,26 @@
+const getRandomNumber = (min, max) => {
+  const lower = Math.ceil(Math.min(Math.abs(min), Math.abs(max)));
+  const upper = Math.floor(Math.max(Math.abs(min), Math.abs(max)));
+  const result = Math.random() * (upper - lower + 1) + lower;
+  return Math.floor(result);
+};
+
+const getRandomFloat = (min, max, digits = 1) => {
+  const lower = Math.min(Math.abs(min), Math.abs(max));
+  const upper = Math.max(Math.abs(min), Math.abs(max));
+  const result = Math.random() * (upper - lower) + lower;
+  return result.toFixed(digits);
+};
+
+const getRandomArrayLength = (elements) => {
+  const arrayLength = getRandomNumber(1, elements.length);
+  const array = [];
+  for(let item = 0; item < arrayLength; item++) {
+    array.push(elements[item]);
+  }
+  return array;
+};
+
+const getRandomArrayElement = (elements) => elements[getRandomNumber(0, elements.length - 1)];
+
+export {getRandomNumber, getRandomFloat, getRandomArrayElement, getRandomArrayLength};
